requirements.txt
flask==3.0.0
selenium>=4.0.0
pytest>=7.0.0
pytest-html>=3.2.0
pyyaml>=6.0
